+++ apps/docs/sidebars.ts
@@ -133,6 +133,11 @@ const sidebars: SidebarsConfig = {
             },
             {
               type: 'doc',
+              id: 'components/inputs/Combobox/combobox',
+              label: 'Combobox (Alpha)',
+            },
+            {
+              type: 'doc',
               id: 'components/inputs/ControlGroup/controlGroup',
               label: 'ControlGroup',
             },
@@ -696,6 +701,7 @@ const sidebars: SidebarsConfig = {
               id: 'components/other/Calendar/calendar',
               label: 'Calendar',
             },
+            { type: 'doc', id: 'components/other/DateInput/date-input', label: 'DateInput' },
             { type: 'doc', id: 'components/other/DatePicker/date-picker', label: 'DatePicker' },
             { type: 'doc', id: 'components/other/DotCount/dot-count', label: 'DotCount' },
             {