+++ packages/mobile/src/cells/__stories__/ListCell.stories.tsx
@@ -879,7 +879,12 @@ const CondensedListCell = () => {
       />
       <ListCell
         multiline
-        description="Long description with multiple lines. This section can be arbitrarily long and occupy many many lines."
+        description={
+          <Text color="fgMuted" font="body">
+            Long description with multiple lines. This section can be arbitrarily long and occupy
+            many many lines.
+          </Text>
+        }
         detail="Detail"
         end={
           <HStack alignItems="center" gap={1}>