+++ packages/web/src/alpha/tabbed-chips/TabbedChips.tsx
@@ -17,6 +17,10 @@ import {
   type TabsProps,
 } from '../../tabs';
 
+const containerCss = css`
+  isolation: isolate;
+`;
+
 const scrollContainerCss = css`
   &::-webkit-scrollbar {
     display: none;
@@ -87,6 +91,11 @@ export type TabbedChipsBaseProps<T extends string = string> = Omit<
    * @default false
    */
   compact?: boolean;
+  /**
+   * X position offset when auto-scrolling to active tab (to avoid active tab being covered by the paddle on the left side, default: 50px)
+   * @default 50
+   */
+  autoScrollOffset?: number;
 };
 
 export type TabbedChipsProps<T extends string = string> = TabbedChipsBaseProps<T> &
@@ -162,13 +171,14 @@ const TabbedChipsComponent = memo(
       compact,
       styles,
       classNames,
+      autoScrollOffset = 50,
       ...accessibilityProps
     }: TabbedChipsProps<T>,
     ref: React.ForwardedRef<HTMLElement | null>,
   ) {
     const [scrollTarget, setScrollTarget] = useState<HTMLElement | null>(null);
     const { scrollRef, isScrollContentOffscreenLeft, isScrollContentOffscreenRight, handleScroll } =
-      useHorizontalScrollToTarget({ activeTarget: scrollTarget, scrollPadding: 50 });
+      useHorizontalScrollToTarget({ activeTarget: scrollTarget, autoScrollOffset });
 
     const handleScrollLeft = useCallback(() => {
       scrollRef?.current?.scrollTo({ left: 0, behavior: 'smooth' });
@@ -190,7 +200,7 @@ const TabbedChipsComponent = memo(
     return (
       <HStack
         alignItems="center"
-        className={classNames?.root}
+        className={cx(containerCss, classNames?.root)}
         position="relative"
         style={styles?.root}
         testID={testID}