+++ packages/common/CHANGELOG.md
@@ -8,6 +8,68 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.32.2 ((12/18/2025, 07:41 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.32.1 (12/17/2025 PST)
+
+#### 🐞 Fixes
+
+- Fix: update react-native export pointer. [[#259](https://github.com/coinbase/cds/pull/259)]
+
+## 8.32.0 ((12/16/2025, 08:20 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.31.5 ((12/16/2025, 03:27 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.31.4 ((12/16/2025, 10:41 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.31.3 ((12/16/2025, 07:55 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.31.2 ((12/15/2025, 01:12 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.31.1 ((12/15/2025, 10:08 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.31.0 ((12/12/2025, 04:30 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.30.1 ((12/12/2025, 3:00 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.30.0 ((12/12/2025, 02:53 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.29.0 ((12/12/2025, 01:12 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.28.2 ((12/12/2025, 06:23 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.28.1 ((12/10/2025, 04:33 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.28.0 ((12/10/2025, 03:41 PM PST))
+
+This is an artificial version bump with no new change.
+
 ## 8.27.4 ((12/7/2025, 11:54 AM PST))
 
 This is an artificial version bump with no new change.