+++ packages/web/CHANGELOG.md
@@ -8,6 +8,92 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.32.2 (12/18/2025 PST)
+
+#### 🐞 Fixes
+
+- Add fuse.js dependency. [[#256](https://github.com/coinbase/cds/pull/256)]
+
+## 8.32.1 ((12/17/2025, 11:31 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.32.0 (12/16/2025 PST)
+
+#### 🚀 Updates
+
+- Add static class names to components.
+
+## 8.31.5 (12/16/2025 PST)
+
+#### 🐞 Fixes
+
+- Fixed caption font-size on web dense theme.
+
+## 8.31.4 (12/16/2025 PST)
+
+#### 🐞 Fixes
+
+- Update date types barrel to export. [[#255](https://github.com/coinbase/cds/pull/255)]
+
+## 8.31.3 (12/16/2025 PST)
+
+#### 🐞 Fixes
+
+- Isolate tabs container so the paddles don't cover elements outside. [[#254](https://github.com/coinbase/cds/pull/254)]
+
+## 8.31.2 ((12/15/2025, 01:12 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.31.1 (12/15/2025 PST)
+
+#### 🐞 Fixes
+
+- Chore: add barrel files to web/mobile components. [[#251](https://github.com/coinbase/cds/pull/251)]
+
+## 8.31.0 (12/12/2025 PST)
+
+#### 🚀 Updates
+
+- Add inputBackground prop to InputStack and TextInput.
+
+## 8.30.1 (12/12/2025 PST)
+
+#### 🐞 Fixes
+
+- Add classNames and styles to sidebar. [[#239](https://github.com/coinbase/cds/pull/239)]
+
+## 8.30.0 (12/12/2025 PST)
+
+#### 🚀 Updates
+
+- Add autoScrollOffset prop to TabbedChips alpha. [[#232](https://github.com/coinbase/cds/pull/232)]
+- Add maxWidth prop to SelectChip alpha [[#232](https://github.com/coinbase/cds/pull/232)] [DX-5072]
+- Add displayValue prop to SelectChip alpha [[#232](https://github.com/coinbase/cds/pull/232)]
+
+## 8.29.0 (12/12/2025 PST)
+
+#### 🚀 Updates
+
+- Combobox component and header prop for Tray. [[#182](https://github.com/coinbase/cds/pull/182)]
+
+## 8.28.2 ((12/12/2025, 06:23 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.28.1 (12/10/2025 PST)
+
+#### 🐞 Fixes
+
+- Fix a11y bug on ModalHeader.
+
+## 8.28.0 (12/10/2025 PST)
+
+#### 🚀 Updates
+
+- Updated ContentCell to support condensed variant. [[#205](https://github.com/coinbase/cds/pull/205)] [DX-5013]
+
 ## 8.27.4 ((12/7/2025, 11:54 AM PST))
 
 This is an artificial version bump with no new change.